worker/requirements.txt
celery
redis
pillow
python-dotenv
cloudinary
requests
PyMuPDF